♻️ refactor [#10.6.20]: 2차 개선 - Polling 아키텍처 개선 및 최적화
🔧 변경 사항:
- **[Hooks]**: [useVisibilityPolling] `web_ui/src/hooks/use-visibility-polling.ts` 커스텀 훅 신규 구현
  - [src/hooks/use-visibility-polling.ts] `web_ui/src/hooks/use-visibility-polling.ts`
  - `inFlight` 플래그로 중복 요청(Race Condition) 원천 차단
  - Page Visibility API 및 SSR(`typeof document`) 검사로 백그라운드 리소스 절약
  - `useRef`를 활용해 [fetchData] `web_ui/src/components/dashboard/sync-monitor.tsx` 의 불필요한 의존성 제거

- **[SyncMonitor]**: 관심사 분리(Separation of Concerns) 적용
  - 복잡한 Polling 로직을 Custom Hook으로 위임하여 컴포넌트 단순화
  - [fetchData] `web_ui/src/components/dashboard/sync-monitor.tsx` 에서 불필요한 `useCallback` 제거

- **[ConflictResolver]**: Dialog 닫힘 핸들러 방어 코드 추가
  - `onOpenChange`가 `false`일 때만 동작하도록 수정

🎯 효과:
- 네트워크 상태와 무관하게 **안정적이고 효율적인 데이터 동기화** 보장
- 컴포넌트 코드 **가독성 및 유지보수성 향상**

📌 Related:
- Issue [#213]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/249#pullrequestreview-3613325961)
